Refactor search: extract controller hook and constants.
Moves search state management, API orchestration, debounced suggestions, and discover data loading into a reusable feature hook with shared constants.
