+++ mobile/app/(tabs)/search.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useCallback, useEffect, useRef } from 'react';
+import React, { useCallback } from 'react';
 import {
   View,
   Text,
@@ -18,34 +18,11 @@ import { LinearGradient } from 'expo-linear-gradient';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
 import { useRouter } from 'expo-router';
 import { StatusBar } from 'expo-status-bar';
-import {
-  getBooks,
-  getSearchSuggestions,
-  getRandomBooks,
-  type Book,
-  type BookStatus,
-  type SearchSuggestion,
-  BooksApiError,
-} from '@/src/services/books';
+import { type Book, type BookStatus } from '@/src/services/books';
 import { getCoverUrl } from '@/src/utils/covers';
 import { BOTTOM_TAB_BAR_HEIGHT } from '@/components/BottomTabBar';
-
-const HOWARD_BLUE = '#003A63';
-const HOWARD_RED = '#E31837';
-const SEARCH_PAGE_SIZE = 20;
-const RECENT_MAX = 10;
-const SUGGEST_DEBOUNCE_MS = 300;
-
-const POPULAR_LIMIT = 4;
-
-const CATEGORIES = [
-  { label: 'Science', subtitle: 'Lab, Research', icon: 'flask' as const, color: '#06b6d4', bg: '#ecfeff' },
-  { label: 'Fiction', subtitle: 'Stories, Novels', icon: 'magic' as const, color: '#8b5cf6', bg: '#f5f3ff' },
-  { label: 'Technology', subtitle: 'AI, Systems', icon: 'microchip' as const, color: HOWARD_BLUE, bg: '#eff6ff' },
-  { label: 'History', subtitle: 'Archives', icon: 'globe' as const, color: '#f59e0b', bg: '#fffbeb' },
-  { label: 'Biography', subtitle: 'Memoirs, Life', icon: 'user' as const, color: '#10b981', bg: '#ecfdf5' },
-  { label: 'Art & Design', subtitle: 'Creative', icon: 'paint-brush' as const, color: HOWARD_RED, bg: '#fff1f2' },
-];
+import { CATEGORIES, HOWARD_BLUE, HOWARD_RED } from '@/src/features/search/constants';
+import { useSearchController } from '@/src/features/search/useSearchController';
 
 function statusLabel(s: BookStatus): string {
   switch (s) {
@@ -107,127 +84,31 @@ function SearchBookRow({
 export default function SearchScreen() {
   const insets = useSafeAreaInsets();
   const router = useRouter();
-  const [query, setQuery] = useState('');
-  const [submittedQuery, setSubmittedQuery] = useState<string | null>(null);
-  const [results, setResults] = useState<Book[]>([]);
-  const [page, setPage] = useState(1);
-  const [totalPages, setTotalPages] = useState(1);
-  const [loading, setLoading] = useState(false);
-  const [loadingMore, setLoadingMore] = useState(false);
-  const [refreshing, setRefreshing] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-  const [recentSearches, setRecentSearches] = useState<string[]>([]);
-  const [suggestions, setSuggestions] = useState<SearchSuggestion[]>([]);
-  const [popularBooks, setPopularBooks] = useState<Book[]>([]);
-  const [popularLoading, setPopularLoading] = useState(true);
-  const [categoryCounts, setCategoryCounts] = useState<Record<string, number>>({});
-  const suggestTimerRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-
-  useEffect(() => {
-    let cancelled = false;
-    setPopularLoading(true);
-    getRandomBooks(POPULAR_LIMIT)
-      .then((books) => { if (!cancelled) setPopularBooks(books); })
-      .catch(() => { if (!cancelled) setPopularBooks([]); })
-      .finally(() => { if (!cancelled) setPopularLoading(false); });
-    return () => { cancelled = true; };
-  }, []);
-
-  useEffect(() => {
-    let cancelled = false;
-    Promise.all(
-      CATEGORIES.map((cat) =>
-        getBooks({ q: cat.label, page: 1, limit: 1, sort: 'title', order: 'asc' })
-          .then((res) => ({ label: cat.label, total: res.pagination?.total ?? 0 }))
-          .catch(() => ({ label: cat.label, total: 0 }))
-      )
-    ).then((results) => {
-      if (cancelled) return;
-      setCategoryCounts(
-        results.reduce<Record<string, number>>((acc, { label, total }) => {
-          acc[label] = total;
-          return acc;
-        }, {})
-      );
-    });
-    return () => { cancelled = true; };
-  }, []);
-
-  const runSearch = useCallback(async (q: string, pageNum: number = 1, isRefresh = false) => {
-    const trimmed = q.trim();
-    if (!trimmed) return;
-    if (pageNum === 1) {
-      setSubmittedQuery(trimmed);
-      if (isRefresh) setRefreshing(true);
-      else setLoading(true);
-    } else {
-      setLoadingMore(true);
-    }
-    setError(null);
-    try {
-      const res = await getBooks({ q: trimmed, page: pageNum, limit: SEARCH_PAGE_SIZE, sort: 'title', order: 'asc' });
-      if (pageNum === 1) {
-        setResults(res.items);
-        setPage(2);
-        setTotalPages(res.pagination?.total_pages ?? 1);
-        setRecentSearches((prev) => {
-          const next = [trimmed, ...prev.filter((x) => x !== trimmed)].slice(0, RECENT_MAX);
-          return next;
-        });
-      } else {
-        setResults((prev) => [...prev, ...res.items]);
-        setPage((p) => p + 1);
-      }
-    } catch (e) {
-      const msg = e instanceof BooksApiError ? e.message : 'Search failed';
-      setError(msg);
-      if (pageNum === 1) setResults([]);
-    } finally {
-      setLoading(false);
-      setLoadingMore(false);
-      setRefreshing(false);
-    }
-  }, []);
+  const {
+    query,
+    setQuery,
+    submittedQuery,
+    results,
+    loading,
+    loadingMore,
+    refreshing,
+    error,
+    recentSearches,
+    suggestions,
+    popularBooks,
+    popularLoading,
+    categoryCounts,
+    runSearch,
+    submit,
+    loadMore,
+    clearSearchHistory,
+    selectSuggestion,
+  } = useSearchController();
 
   const handleSubmit = useCallback(() => {
     Keyboard.dismiss();
-    const trimmed = query.trim();
-    if (trimmed) runSearch(trimmed, 1);
-  }, [query, runSearch]);
-
-  useEffect(() => {
-    const t = query.trim();
-    if (t.length < 2) {
-      setSuggestions([]);
-      return;
-    }
-    if (suggestTimerRef.current) clearTimeout(suggestTimerRef.current);
-    suggestTimerRef.current = setTimeout(() => {
-      getSearchSuggestions(t, 10)
-        .then(setSuggestions)
-        .catch(() => setSuggestions([]));
-      suggestTimerRef.current = null;
-    }, SUGGEST_DEBOUNCE_MS);
-    return () => {
-      if (suggestTimerRef.current) clearTimeout(suggestTimerRef.current);
-    };
-  }, [query]);
-
-  const loadMore = useCallback(() => {
-    if (!submittedQuery || loadingMore || page > totalPages) return;
-    runSearch(submittedQuery, page);
-  }, [submittedQuery, page, totalPages, loadingMore, runSearch]);
-
-  const clearSearchHistory = useCallback(() => {
-    setRecentSearches([]);
-    setQuery('');
-    setSubmittedQuery(null);
-    setResults([]);
-    setError(null);
-    setPage(1);
-    setTotalPages(1);
-    setSuggestions([]);
-  }, []);
+    submit();
+  }, [submit]);
 
   const showResults = Boolean(submittedQuery?.trim());
   const showDiscover = !showResults;
@@ -263,9 +144,7 @@ export default function SearchScreen() {
                 key={`${s.label}-${s.type}-${i}`}
                 style={styles.suggestItem}
                 onPress={() => {
-                  setQuery(s.label);
-                  setSuggestions([]);
-                  runSearch(s.label, 1);
+                  selectSuggestion(s);
                 }}
               >
                 <FontAwesome name={s.type === 'author' ? 'user' : 'book'} size={14} color="#64748b" />
